+++ public/js/i18n/da_DK.js
@@ -227,6 +227,8 @@ gc2i18n = {
         "This will create one new composite layer with all layers in the GGIS project. Blending mode from QGIS will be used. This is great for creating a base layer. The new layer will take name from the qgs-file and be placed in the current schema": "Dette vil oprette et nyt composite lag med alle lag fra QGIS projektet. Blending mode vil blive anvendt. Dette kan bruges til at skabe et baggrundskort. Det nye lag vil tage navn fra qgs-filen og blive placeret i nuværende schema.",
 
         "Append" : "Tilføj",
-        "Write tag names into the field 'Tags'. Finish the tag name with an 'Enter' key stroke. If 'Append' is checked, new tags are appended to existing one. Append is only available when multiple layers are selected.": "Skriv tag-navne i feltet 'Tags'. Afslut tag-navnet ved at taste 'Enter'. Hvis 'Tilføj' er tjekket af, tilføjes nye tags til eksisterende. 'Tilføj' er kun tilgængelig, når flere lag er valgt."
+        "Write tag names into the field 'Tags'. Finish the tag name with an 'Enter' key stroke. If 'Append' is checked, new tags are appended to existing one. Append is only available when multiple layers are selected.": "Skriv tag-navne i feltet 'Tags'. Afslut tag-navnet ved at taste 'Enter'. Hvis 'Tilføj' er tjekket af, tilføjes nye tags til eksisterende. 'Tilføj' er kun tilgængelig, når flere lag er valgt.",
+
+        "Change between WMS and Tile Cache": "Skift mellem WMS og Tile Cache"
     }
 };