+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
 ## [Unreleased]
+### Added
 - Phpfastcache added to speed up Meta API.
 - MapServer/MapCache now exposes layers as Mapbox Vector tiles. In MapServer just use `format=mvt` and in MapCache MVT layers are prefixed with `.mvt` like public.foo.mvt.
+- Change between single WMS tile and MapCache tiles in layer tree.
 
 ### Fixed
 - Limit of 100 classes in sorting algorithms is increased.